More consistent rectangle data handling

+++ facerec.h
@@ -19,7 +19,7 @@ typedef struct facerec {
 
 typedef struct faceret {
 	int num_faces;
-	long* rectangles;
+	int32_t* rectangles;
 	float* descriptors;
 	const char* err_str;
 	err_code err_code;
